==> analysis_pipeline/01_ASGS_Lookup.Rmd <==
---
title: "01_ASGS_Lookup"
author: "Calum Nicholson"
date: "03/04/2021"
output: html_document
---

```{r setup, include=FALSE, cache = FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(dplyr)
```

## The SA2 Coding index 

Allows SA2 areas to be looked up using with locality (suburb), Postcode and State.
``` {r}
# Import SA2 coding index
SA2_locality <- read.csv('./input/2019 Locality to 2016 SA2 Coding Index.csv')

# filter for NSW
SA2_locality_NSW <- SA2_locality %>%
                    filter(STATE == "NSW" | STATE == "ACT")
# Get distinct SA2 names
SA2_locality_names <- SA2_locality_NSW %>%
                      distinct(SA2_NAME)

```

## Check how the coding index compared with the total list of SA2 areas

``` {r}
# Load the list of ASGS SA2 area names and codes
ASGS_aus <- read.csv('./input/SA2_2016_AUST.csv') %>%
            select(STATE_NAME_2016, SA2_MAINCODE_2016, SA2_5DIGITCODE_2016, SA2_NAME_2016,
                   SA3_CODE_2016, SA3_NAME_2016, SA4_CODE_2016, SA4_NAME_2016, GCCSA_NAME_2016)

# filter for NSW and select the column with SA2 names
ASGS_NSW_ACT <- ASGS_aus %>%
            filter(STATE_NAME_2016 == "New South Wales" | STATE_NAME_2016 == "Australian Capital Territory")
```


``` {r}
#Get missing records
missing.records <- anti_join(ASGS_NSW_ACT, SA2_locality_names, by = c("SA2_NAME_2016" = "SA2_NAME"))

missing.records
```


A number of SA2 areas are missing from the localities index but looking at the documentation (Dissertation/data/ASGS/asgs2016codingindexes/2017 Locality to 2016 SA2 Coding Index Documentation.docx) we can see the methods used to create the coding indexes:

_This means where a locality is covered by two or more SA2s the locality has been allocated to the SA2 where most of its population are located according to the underlying grid.  This ensures a specific gazetted locality/alias/extra locality entry is represented only once in the file and is allocated wholly to a specific SA2._

This method may mean that some SA2 areas have been excluded because the locality has been placed into the area where most of its population is.

``` {r}
# join the SA3, SA4 and GCC area codes to the look up table
ASGS_lookup <- left_join(SA2_locality_NSW, ASGS_aus, by = c("SA2_MAINCODE" = "SA2_MAINCODE_2016")) %>%
                    select(-LOCALITY_ID, -LOCALITY_TYPE, -STATE_NAME_2016, -SA2_NAME_2016) %>%
                    rename("GCC_NAME" = GCCSA_NAME_2016,
                           "SA2_5DIGIT" = SA2_5DIGITCODE_2016,
                           "SA3_CODE" = SA3_CODE_2016,
                           "SA3_NAME" = SA3_NAME_2016,
                           "SA4_CODE" = SA4_CODE_2016,
                           "SA4_NAME" = SA4_NAME_2016)
```

``` {r}
# check for missingness in the ASGS lookup
sapply(ASGS_lookup, function(x) sum(is.na(x)))

ASGS_lookup %>% filter(is.na(POSTCODE))
```

``` {r}
# Better select relevant columns and choose better column names
ASGS_NSW_ACT <- ASGS_NSW_ACT %>%
                select(-STATE_NAME_2016, -SA2_MAINCODE_2016) %>%
                    rename("GCC_NAME" = GCCSA_NAME_2016,
                           "SA2_NAME" = SA2_NAME_2016,
                           "SA2_5DIGIT" = SA2_5DIGITCODE_2016,
                           "SA3_CODE" = SA3_CODE_2016,
                           "SA3_NAME" = SA3_NAME_2016,
                           "SA4_CODE" = SA4_CODE_2016,
                           "SA4_NAME" = SA4_NAME_2016)


```

``` {r}
# write the filtered lookup table to the output folder
 write.csv(ASGS_NSW_ACT, "./output/ASGS_areas.csv")
 write.csv(ASGS_lookup, "./output/ASGS_lookup.csv")

```
